requirements.txt
streamlit
sqlalchemy
psycopg2-binary
google-generativeai
python-dotenv
pandas
plotly
openai
